% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\title{Counts and proportions}
\usage{
counts(formula, data, ..., tidy = TRUE)
}
\arguments{
\item{level}{set to, e.g. 0.95 for the confidence intervals
to be generated.}
}
\description{
Counts and proportions
}
