% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election.R
\name{election}
\alias{election}
\title{Simulate the outcome of an election based on polling data}
\usage{
election(formula, data, N = 1000, weights = NULL, turnout = 1)
}
\arguments{
\item{formula}{\itemize{
\item the model of individual vote outcome
}}

\item{data}{\itemize{
\item the data table containing the polling sample
}}

\item{N}{\itemize{
\item the size of the electorate
}}

\item{weights}{\itemize{
\item one-sided formula giving variable to be used for weights
}}

\item{turnout}{\itemize{
\item a fraction (e.g. 0.60) or a one-sided formula giving the name
of the variable expressing turnout probability for each voter in the poll. NOT YET IMPLEMENTED.
}}
}
\description{
Takes a data table with polling information about A/B preferences and
covariates for each polled individual. Fit's the specified model
of outcome and runs 100 simulated elections with a total electorate size of N
(as opposed to the poll size n).
}
\examples{
D <- data.frame(response = c(0,0,1,1,1,1), weight = 1:6, sex = c("M", "F", "M", "F", "F", "F"))
election(response ~ 1, D, N = 1000)
}
