% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_trials.R
\name{gf_do}
\alias{gf_do}
\title{Iterate a graphic}
\usage{
gf_do(formula, what, ntrials = 3, data, start = NULL, ...)
}
\arguments{
\item{data}{a data table or, more usually, an expression that produces
a data table, e.g. \code{mosaic::resample(CPS85)}}
}
\description{
Iterate a graphic
}
\details{
It makes sense to use this only for graphics that do a
substantial calculation on the data.
}
\examples{
gf_do(wage ~ sector + color:sex + alpha:.1, gf_boxplot,
  data = mosaic::resample(CPS85), ntrials = 10)
gf_do(~ wage + color:sex + alpha:.05, gf_freqpoly,
  data = mosaic::resample(CPS85), ntrials = 10)
gf_do(wage ~ sex + alpha:.05, gf_violin,
  data = mosaic::resample(CPS85), ntrials = 10)
gf_do(~ wage  + fill:sex + alpha:.05, gf_histogram,
  data = mosaic::resample(CPS85), ntrials = 50) \%>\%
  gf_facet_wrap(~sex)
}
