% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest.R
\name{htest}
\alias{htest}
\title{Hypothesis tests}
\usage{
htest(formula, data, ..., test = c("t", "chisq", "exact", "goodness",
  "independence", "coefficients", "anova"), method = c("linear", "logistic"),
  prob_model = NULL)
}
\arguments{
\item{formula}{a formula expression the relationship among the variable}

\item{data}{a data table containing the variables used in the formula}

\item{...}{other arguments for the particular test}

\item{test}{character string naming the particular flavor of test to conduct:
"t", "chisq", "exact" (Fisher exact test), "coefficients" (as in regression),
"anova"}

\item{method}{for regression, whether to do "linear" regression (default) or logistic regression}
}
\description{
\code{htest} carries out a variety of hypothesis tests commonly encountered in an
introductory statistics course.
}
