% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtally.R
\name{mtally}
\alias{mtally}
\title{mtally}
\usage{
mtally(.data, ...)
}
\arguments{
\item{.data}{placeholder for a piped-in data set.}

\item{...}{the formula and other arguments for mosaic::tally}
}
\description{
A function to disambiguate between dplyr::tally and mosaic::tally.
mtally works like mosaic::tally, producing a table-class rather than a data table.
The table-class output is better suited as a data presentation to people,
and is also suited to functions such as chisq.test() or fisher.test()
mtally() is slightly different from mosaic::tally() in accepting a piped
data frame as input
}
\examples{
mtally(mtcars, ~ cyl)
}
\seealso{
tally in the mosaic package
}
