% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{do}
\alias{reexports}
\alias{ntiles}
\alias{reexports}
\alias{gmodel}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{mosaic}{\code{\link[mosaic]{do}}, \code{\link[mosaic]{ntiles}}}

  \item{statisticalModeling}{\code{\link[statisticalModeling]{gmodel}}}
}}

