% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstats.R
\name{qstats}
\alias{qstats}
\title{Calculate basic statistics on a quantitative variable}
\usage{
qstats(formula, data, ..., .level = 0.95, .wide = FALSE)
}
\description{
Calculate basic statistics on a quantitative variable
}
