% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_n.R
\name{resample_n}
\alias{resample_n}
\title{Resample in the dplyr style}
\usage{
resample_n(tbl, size = NULL, weight = NULL, .env = parent.frame())
}
\description{
This is a simple wrapper on \code{dplyr::sample_n`` that sets}replace=TRUE`
}
